app/requirements.txt
fastapi==0.65.2
loguru==0.5.3
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.17.1